+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-07-12
+### Changed
+- added css variables to material-button for font size and icon size
+- added setData() method to material-dropdown to dynamically add data
+- added missing events to documentation
+
 ## [1.1.5] - 2019-06-18
 ### Added
 - added material-card component